+++ docs/changes.rst
@@ -7,6 +7,7 @@ OpenREM version history
 --------------------
 * :issue:`911`  Charts: fixed issue with chart data sorting and added label wrap option
 * :issue:`909`  Code quality: all model imports absolute
+* :issue:`908`  Imports: enabled Device Observer UID to be ignored for specific equipment models when creating display name data during RDSR import
 * :issue:`906`  Charts: upgraded Plotly library to latest version
 * :issue:`905`  Imports: fixed filter extraction code not working for Siemens Multix DX
 * :issue:`904`  Testing: bumped Python image from 3.6 to 3.8
@@ -17,6 +18,7 @@ OpenREM version history
 * :issue:`893`  Charts: fixed issue with over-time charts with many sub-plots failing to plot correctly
 * :issue:`888`  SkinDose: added option to support skin dose calculations for listed systems only
 * :issue:`886`  Code quality: addressed some SonarCloud issues
+* :issue:`882`  SkinDose: added percentage of exposures that interact with phantom
 * :issue:`881`  Charts: add option to remove multiple and trailing whitespace in category names
 * :issue:`879`  Charts: fixed sorting of fluoroscopy charts when split by physician
 * :issue:`877`  Charts: added acquisition type restrictions to acquisition-level CT charts
@@ -53,7 +55,7 @@ OpenREM version history
 * :issue:`830`  Charts: fixed incorrect histogram data in charts
 * :issue:`829`  Installation: added docs for Docker install on computer without internet access
 * :issue:`828`  Docker: enabled email configuration to work
-* :issue:`827`  SkinDose: Made SkinDose results available in OpenREM and made alert triggering possible
+* :issue:`827`  SkinDose: made SkinDose results available in OpenREM and made alert triggering possible
 * :issue:`826`  Code quality: split views to make more manageable and testable
 * :issue:`824`  DICOM Networking: enabled declaration and testing of Orthanc Store SCP in Docker
 * :issue:`821`  Code quality: fixed literal comparisons Docker was complaining about